Wire::fSignal now is a vector of floats
Former-commit-id: d44fd6b16ab6d32a274102b3c8be68665eed13eb

+++ HitFinder/FFTHitFinder.cxx
@@ -104,8 +104,8 @@ namespace hit{
       startTimes.clear();
       maxTimes.clear();
       endTimes.clear();
-      std::vector<double> signal(wire->Signal());
-      std::vector<double>::iterator timeIter;   // iterator for time bins
+      std::vector<float> signal(wire->Signal());
+      std::vector<float>::iterator timeIter;   // iterator for time bins
       time          = 0;
       minTimeHolder = 0;
       maxFound      = false;
